feat: not antlr-dependant enrichWith function
