+++ server/src/visitors/YmlEnumVisitor.ts
@@ -1,10 +1,10 @@
 import { YmlCompletionItemsProvider } from '../completion/YmlCompletionItemsProvider';
 import { YmlDefinitionProvider } from '../definitions';
 import { EnumElementContext, YenumContext } from '../grammar/YmlParser';
-import { connection } from '../server';
 import { YmlEnum } from '../yml-objects/YmlEnum';
 import { YmlEnumMember } from '../yml-objects/YmlEnumMember';
 import { YmlBaseVisitor } from './YmlBaseVisitor';
+import { getDocumentation, getType } from './YmlVisitorHelper';
 
 export class YmlEnumVisitor extends YmlBaseVisitor {
     /**
@@ -37,7 +37,9 @@ export class YmlEnumVisitor extends YmlBaseVisitor {
 
     public visitEnumElement(node: EnumElementContext): void {
         const enumMember = new YmlEnumMember(`${this.enumName}::${node.ymlId().text}`, this.uri);
-        enumMember.enrichWith(node.field(), connection, this.enumName, this.enumName);
+        const doc = getDocumentation(node.field());
+        const type = getType(node.field(), this.enumName);
+        enumMember.enrichWith(doc, type, this.enumName);
         this.completionProvider.addCompletionItem(enumMember);
         enumMember.setDefinitionLocation(node.start, node.stop, this.uri);
         this.definitions.addDefinition(enumMember);
